WCM-435: move IConverter into its own module to reuse it in models
- avoiding circular imports
